Move URL inspection tests into existing test layout
